feat: taxonomia fechada, score ponderado e exportacao CSV (Sprint A)
- Novo modulo core/taxonomy.py com CATEGORIAS, SENTIMENTOS, CRITICIDADES
- validar_classificacao() normaliza saida da IA contra a taxonomia fechada
- calcular_score_final() recalcula score como media ponderada (pesos iguais 25%)
- Prompt instrui a IA a escolher categoria/sentimento/criticidade da lista
- Campo nota muda de Integer para Float (preserva 7.5)
- Migration automatica converte coluna Nota para double precision
- GET /atendimentos/export retorna CSV (delimitador ; e BOM para Excel)
- Dashboard agrupa notas float em buckets inteiros somando corretamente

+++ app/schemas.py
@@ -17,7 +17,7 @@ class MensagemOut(_ORMModel):
 
 class AvaliacaoOut(_ORMModel):
     id: int
-    nota: Optional[int] = None
+    nota: Optional[float] = None
     comentario: Optional[str] = None
     avaliado_em: Optional[datetime] = None
     aprovado_como_exemplo: bool = False
